requirements.txt
kittyfuzzer==0.7.1
